+++ src/librawspeed/tiff/TiffEntry.cpp
@@ -261,7 +261,7 @@ float TiffEntry::getFloat(uint32_t index) const {
   switch (type) {
     using enum TiffDataType;
   case DOUBLE:
-    return data.peek<double>(index);
+    return implicit_cast<float>(data.peek<double>(index));
   case FLOAT:
     return data.peek<float>(index);
   case LONG:
